v1.0.1 - initial release
v1.0.0 was technically released beforehand, but I did not commit this,
and that file was very quickly updated to fix the web release, to add an
additional contribution to the menu, etc.
This was finished a few days ago, but I burnt myself out, so I took a
rest before making this commit.

v1.0.0 added CPUs to the game, alongside Rouge, Omega, Zavok and Doom's
Eye.
It adds the main menus, alongside the winning of races.
Iron Jungle was finished, with features like the Minimap added, and
probably more I can't remember.
Some Trans Pride was added into Westopolis, where one of the Chaos
Emeralds was planned to be.
Probably a shit ton more was done on the final day, I do not have much
memory of it.